+++ src/GEOComp.tsx
@@ -1,7 +1,7 @@
+import { useState,useEffect,useCallback } from 'react';
 import {
   UICompBuilder,
   NameConfig,
-  BoolControl,
   stringSimpleControl,
   JSONObjectControl,
   NumberControl,
@@ -12,16 +12,14 @@ import {
   eventHandlerControl,
   styleControl,
   jsonObjectExposingStateControl,
-  stringExposingStateControl,
   AutoHeightControl,
+  arrayStringExposingStateControl,
 } from "lowcoder-sdk";
 import styles from "./styles.module.css";
 import { trans } from "./i18n/comps";
 import { Geo } from "./vendors";
 import { useResizeDetector } from "react-resize-detector";
-import { useState } from "react";
 import {version} from '../package.json';
-import { ObjectEvent } from "ol/Object";
 
 
 export const CompStyles = [
@@ -63,93 +61,159 @@ export const CompStyles = [
 ] as const;
 
 
-let GEOComp = (function () {
-  //Function to prevent unneeded redraws
-  var _skipRedraw = false
-  const skipRedraw = function(){
-    var ret = _skipRedraw
-    _skipRedraw = false
-    return ret
-  }
-
+var GEOComp = (function () {
+  /* By setting the following items within default you can control behavior
+     center:[] will disable automatich centering
+     debug: true will show eventlog to console
+     buttons: { //All buttons are shown by default
+        menu: false,
+        zoom: false,
+        point : false,
+        line: false,
+        polygon: false,
+        redo: false,
+        undo: false,
+        save:false,
+        scale:false,
+        fullscreen:false,
+        layers:false,
+        swipeVertical: false,
+        swipeHorizontal: false,
+        timeline: false,
+        location:false,
+        tracker:false,
+        rotateNorth: false,
+      }
+  */
+  const events = [
+    {
+      label: "onDraw",
+      value: "draw",
+      description: "Triggers when drawLayer data changes",
+    },
+    {
+      label: "onTracker",
+      value: "tracker",
+      description: "Triggers when trackerLayer data changes",
+    },
+    {
+      label: "onLoad",
+      value: "loaded",
+      description: "Triggers when GEO data is loaded",
+    },
+    {
+      label: "onSwipe",
+      value: "swipe",
+      description: "Triggers when on swipe events",
+    },
+    {
+      label: "onClick",
+      value: "click",
+      description: "Triggers when there is a click within the viewer",
+    },
+    {
+      label: "onZoom",
+      value: "zoom",
+      description: "Triggers when there is a zoom change within the viewer",
+    },
+    {
+      label: "onBbox",
+      value: "bbox",
+      description: "Triggers when there is a bbox change",
+    },
+    {
+      label: "onOther",
+      value: "other",
+      description: "Triggers when there is no special event handler is triggered",
+    },
+  ];
   const childrenMap = {
     autoHeight: withDefault(AutoHeightControl, "fixed"),
     styles: styleControl(CompStyles),
-    defaults: withDefault(JSONObjectControl,"{zoom:10,maxZoom:30}"),
+    defaults: withDefault(JSONObjectControl,`{
+      zoom:10,
+      maxZoom:30,
+      menuTitle: "Menu",
+      menuContent: "No Content",
+      debug:true
+    }`),
     center: ArrayControl,
+    layers: withDefault(ArrayControl,`[
+      {
+        type : 'xyz',
+        source : {
+          url :  'https://service.pdok.nl/brt/achtergrondkaart/wmts/v2_0/standaard/EPSG:3857/{z}/{x}/{y}.png'
+        }
+      }
+    ]`),
     zoom: NumberControl,
     maxZoom: NumberControl,
     rotation: NumberControl,
-    pitch: NumberControl,
-    geoJson : jsonObjectExposingStateControl("geoJson"),
+    bbox: arrayStringExposingStateControl([]),
+    menuTitle: stringSimpleControl(""),
+    menuContent: stringSimpleControl(""),
+    drawLayer : jsonObjectExposingStateControl("drawLayer",{"type":"FeatureCollection","features":[{"type":"Feature","geometry":{"type":"LineString","coordinates":[[514138.9757700867,6865494.523372142],[528910.431486197,6856739.497812072]]},"properties":null}]}),
+    trackerLayer : jsonObjectExposingStateControl("trackerLayer",{"type":"FeatureCollection","features":[{"type":"Feature","geometry":{"type":"LineString","coordinates":[[514138.9757700867,6865494.523372142],[528910.431486197,6856739.497812072]]},"properties":null}]}),
     event : jsonObjectExposingStateControl("event"),
-    showLogo : withDefault(BoolControl,true),
-    onEvent: eventHandlerControl([
-      {
-        label: "onChange",
-        value: "change",
-        description: "Triggers when GeoJson data changes",
-      },
-      {
-        label: "onLoadEnd",
-        value: "loadend",
-        description: "Triggers when GEO data is loaded",
-      },
-      {
-        label: "onClick",
-        value: "click",
-        description: "Triggers when there is a click within the viewer",
-      },
-      {
-        label: "onZoom",
-        value: "zoom",
-        description: "Triggers when there is a zoom change within the viewer",
-      },
-    ] as const),
+    buttons: withDefault(JSONObjectControl,""),
+    onEvent: eventHandlerControl(events),
   };
+
+  //ignoreUpdate function
+  const _ignoreUpdate : any = {}
+  const setIgnoreUpdate = function(name : string){
+    _ignoreUpdate[name]=true
+  }
+  const ignoreUpdate = function(name : string){
+    var ret = _ignoreUpdate[name] || false
+    _ignoreUpdate[name] = false
+    return ret
+  }
    
   return new UICompBuilder(childrenMap, (props: {
     onEvent: any;
     styles: { backgroundColor: any; border: any; radius: any; borderWidth: any; 
               margin: any; padding: any; textSize: any; };
     center : any;
-    pitch : number;
     zoom : number;
     maxZoom: number;
     rotation: number;
-    geoJson: any;
+    drawLayer: any;
+    layers: any;
+    bbox: any;
+    trackerLayer:any;
     event : any;
     defaults: any;
-    /*
-    values: object | null | undefined;
-    svgDownload: boolean;
-    imageName : string;
-    designer: boolean;
-    */
-    showLogo : boolean;
+    buttons: any;
+    menuTitle:string;
+    menuContent:string;
     autoHeight: boolean;
   }) => {
-  const handleDataChange = (json: string) => {
-    //TODO: Set output variable  props.geoJson.onChange(json);
-    _skipRedraw = true //We should not redraw the component
-    props.geoJson.onChange(json);
-    props.onEvent("change");
-  };
-  const handleLoadEnd= (event : object) =>{
-    props.event = event
-    props.onEvent("loadend");
-  };
-  const handleClick = (event : object,notify: any) =>{
-    props.event = event
-    if (props.onEvent("click")) {
-      notify.show("Clik")
-    }
-  }
 
-  const handleZoom= (event: ObjectEvent,newValue : number) =>{
-      props.event = Object.assign({},event,{newValue})
-      props.onEvent("zoom")
-  }
+  //The event handler will also sent the event value to use
+  const handleEvent = useCallback((name : string, eventObj : object,notify: any)=>{
+    props.event.onChange(Object.assign(props.event.value || {},{
+      [name] : eventObj,
+      current   : name
+    }))
+    var n = name.split(":")[0]
+    var eventName = "other"
+    events.forEach((k)=>{if (k.value==n) {eventName=k.value}})
+    switch (eventName){
+      case 'draw': 
+         setIgnoreUpdate('drawLayer')
+         props.drawLayer.onChange(eventObj); 
+         break; //Set the drawLayer object
+      case 'bbox': 
+         props.bbox.onChange(eventObj)
+         break;
+      //case 'tracker': props.trackerLayer.onChange(eventObj); break; //Set the drawLayer object
+    }
+    props.onEvent(eventName,eventObj);
+    if (props.defaults && props.defaults.debug===true)
+       console.log("handleEvent",eventName,props.event.value)
+  },[props.onEvent,props.event]);
+
   const [dimensions, setDimensions] = useState({ width: 480, height: 415 });
   const { width, height, ref: conRef } = useResizeDetector({onResize: () =>{
     const container = conRef.current;
@@ -185,20 +249,19 @@ let GEOComp = (function () {
     }}>
       <Geo
         center={ props.center}
-        geoJson={props.geoJson.value}
+        drawLayer={props.drawLayer.value}
         zoom={props.zoom }
         maxZoom={props.maxZoom}
-        pitch={props.pitch }
         rotation={props.rotation}
         height={dimensions.height}
         width={dimensions.width}
-        showLogo={props.showLogo}
-        onDataChange={handleDataChange}
-        onLoadEnd={handleLoadEnd}
-        onZoom={handleZoom}
-        onClick={handleClick}
-        skipRedraw={skipRedraw}
+        buttons={props.buttons}
+        menuContent={props.menuContent}
+        menuTitle={props.menuTitle}
         defaults={props.defaults}
+        layers={props.layers}
+        onEvent={handleEvent}
+        ignoreUpdate={ignoreUpdate}
       />
     </div>
   );
@@ -206,15 +269,20 @@ let GEOComp = (function () {
 .setPropertyViewFn((children: any) => {
   return (
     <>
-      <Section name="Config">
-        {children.geoJson.propertyView({ label: "geoJson" })}
+      <Section name="Map Layers">
+      {children.layers.propertyView({ label: "layers" })}
+      {children.drawLayer.propertyView({ label: "draw" })}
+      </Section>
+      <Section name="View">
         {children.center.propertyView({ label: "center" })}
         {children.zoom.propertyView({ label: "zoom" })}
         {children.maxZoom.propertyView({ label: "maxZoom" })}
-        {children.pitch.propertyView({ label: "pitch" })}
         {children.rotation.propertyView({ label: "rotation" })}
-        <span>Hide <b>logo</b> only if you are entitled</span>
-        {children.showLogo.propertyView({ label: "Show logo" })}
+        {children.buttons.propertyView({ label: "buttons" })}
+      </Section>
+      <Section name="Menu">
+        {children.menuTitle.propertyView({ label: "title" })}
+        {children.menuContent.propertyView({ label: "content" })}
       </Section>
       <Section name="Interaction">
         {children.onEvent.propertyView()}
@@ -222,7 +290,7 @@ let GEOComp = (function () {
       <Section name="Styles">
         {children.styles.getPropertyView()}
       </Section>
-      <Section name="JSON Controled">
+      <Section name="Advanced">
         {children.defaults.propertyView({ label: "defaults" })}
       </Section>
       <div >
@@ -241,13 +309,8 @@ GEOComp = class extends GEOComp {
 };
 
 export default withExposingConfigs(GEOComp, [
-  new NameConfig("defaults", trans("component.defaults")),
-  new NameConfig("center", trans("component.center")),
-  new NameConfig("zoom", trans("component.zoom")),
-  new NameConfig("maxZoom", trans("component.maxZoom")),
-  new NameConfig("rotation", trans("component.rotation")),
-  new NameConfig("pitch", trans("component.pitch")),
-  new NameConfig("geoJson", trans("component.geoJson")),
+  new NameConfig("drawLayer", trans("component.drawLayer")),
+  new NameConfig("trackerLayer", trans("component.trackerLayer")),
   new NameConfig("event", trans("component.event")),
-  new NameConfig("showLogo", trans("component.showLogo")),
+  new NameConfig("bbox", trans("component.bbox")),
 ]);