+++ src/vendors/Geo.jsx
@@ -4,23 +4,22 @@ import 'ol-ext/dist/ol-ext.css';
 import "@fortawesome/fontawesome-free/css/all.css"
 import "./styles.css";
 
-import React, { useState } from 'react';
+import { useState,useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types'
+//React spinner
 import {RingLoader} from 'react-spinners'
 
 //The real GEO OpenLayers packages
 import {Map, View} from 'ol/index';
-import {Tile as TileLayer} from 'ol/layer';
-import {XYZ} from 'ol/source';
+import { Vector as VectorLayer} from 'ol/layer';
+import { TileWMS, Vector as VectorSource} from 'ol/source';
+import TileLayer from 'ol/layer/WebGLTile.js';
 import {LineString,Polygon} from 'ol/geom';
-//import ZoomSlider from 'ol/control/ZoomSlider.js';
-import {fromLonLat} from 'ol/proj';
+import { fromLonLat, transformExtent } from 'ol/proj';
 import {FullScreen, Zoom } from 'ol/control';
-import {Style, Stroke} from 'ol/style';
-import GeoJSON from 'ol/format/GeoJSON';
-import VectorLayer from 'ol/layer/Vector';
-import VectorSource from 'ol/source/Vector';
-import OSM from 'ol/source/OSM'
+import {GeoJSON}  from 'ol/format';
+
+//Openlayer Extend imports
 import GeolocationBar from 'ol-ext/control/GeolocationBar'
 import GeolocationButton from 'ol-ext/control/GeolocationButton'
 import CanvasScaleLine from 'ol-ext/control/CanvasScaleLine'
@@ -31,42 +30,54 @@ import Toggle from 'ol-ext/control/Toggle'
 import Select from 'ol-ext/control/Select'
 import Overlay from 'ol-ext/control/Overlay'
 import Timeline from 'ol-ext/control/Timeline'
+import Swipe from 'ol-ext/control/Swipe'
 import {Draw} from 'ol/interaction'
 import UndoRedo from 'ol-ext/interaction/UndoRedo'
 
+///Local import
+import RotateNorthControl from './RotateNorthControl'
+import {createLayer} from './helpers/Layers'
+import {lightStroke,darkStroke,geoJsonStyle} from './helpers/Styles'
+
 
 function Geo(props) {
-  const [geoRef, setGeoRef] = React.useState();
-  const [geoLoc,setGeoLoc] = React.useState();
-  const geoId = Math.random().toString(16).slice(2);
+  const [geoRef, setGeoRef] = useState();
+  const [geoLoc,setGeoLoc] = useState();
+  const [map, setMap] = useState();
+  const [geoId] = useState(Math.random().toString(16).slice(2));
+  //Global notification item
+  const [notification] = useState(new Notification({}))
+  // Vector layer for drawing
+  const [drawVector] = useState(new VectorLayer({
+      name: 'draw',
+      source: new VectorSource(),
+      style: geoJsonStyle
+  }))
+  // Vector layer for the tracker
+  const [trackerVector] = useState(new VectorLayer({
+    name: 'tracker',
+    source: new VectorSource(),
+  }))
 
-  const variants = [
-    'standaard',
-		'pastel',
-		'grijs',
-		'water',
-  ];
+  //Function to check if updating of a variable is allowed
+  const allowUpdate = function(name){
+    return !(props.ignoreUpdate && props.ignoreUpdate(name))
+  }
 
-  var varaint = variants[0]
-  var lightStroke = new Style({
-    stroke: new Stroke({
-      color: [255, 255, 255, 0.6],
-      width: 2,
-      lineDash: [4,8],
-      lineDashOffset: 6
-    })
-  });
+  const fireEvent = function(name ,eventObject ){
+    if (props.onEvent) {
+      props.onEvent(name,eventObject || {},notification)
+    }
+  }
+
+  //All buttons are shown by default
+  const showButton = function (name ) {
+    return ((props.buttons && props.buttons[name]===false) ||
+      (props.defaults && props.defaults.buttons && props.defaults.buttons[name]===false)) ? false : true
+  }
   
-  var darkStroke = new Style({
-    stroke: new Stroke({
-      color: [0, 0, 0, 0.6],
-      width: 2,
-      lineDash: [4,8]
-    })
-  });
-
-   const useGeoRef = React.useCallback(ref => {
-    //Fetch the geolocation based on browser or ip when center is not set
+  //Fetch the geolocation based on browser or ip when center is not set
+  const elementRef = useCallback(ref => {
     if ((props.center && props.center.length==2) || (props.defaults && props.defaults.center)) {
       setGeoRef(ref);
     } else {
@@ -92,105 +103,99 @@ function Geo(props) {
     }
   }, []);
 
-  React.useEffect(() => {
-    if (geoRef && !props.skipRedraw()) {
-      //console.log("Redraw")
-      // Rebuild the GEOL7
+  //Configuration of Map component, changing watch props will rebuild map object
+  useEffect(() => {
+    if (geoRef) {
       geoRef.innerHTML = "<div id='GEO_"+ geoId+ "' style='height:"+props.height+"px;position:relative'></div>"
-      //const geoCanvas = document.getElementById("GEO_"+geoId) 
 
-      //The base layer containg the streetmap
-      var baseLayer = new TileLayer({
-        source: new XYZ({
-          url: //'https://service.pdok.nl/hwh/luchtfotorgb/wmts/v1_0/Actueel_ortho25/EPSG:3857/{z}/{x}/{y}.jpeg'
-          'https://service.pdok.nl/brt/achtergrondkaart/wmts/v2_0/standaard/EPSG:3857/{z}/{x}/{y}.png'
-        })
-      })
-
-      var osm = new TileLayer({
-        title: "OSM",
-        source: new OSM(),
-      })
-
-      const geoJson = new VectorLayer({
-        background: '#1a2b39',
-        source: new VectorSource({
-          url: 'https://openlayers.org/data/vector/ecoregions.json',
-          format: new GeoJSON(),
-        }),
-        style: {
-          'fill-color': ['string', ['get', 'COLOR'], '#eee'],
-        },
-      });
-
-      // New vector layer
-      var vector = new VectorLayer({
-        name: 'tracker',
-        source: new VectorSource(),
-        style: [lightStroke, darkStroke],
-
-      })
-      
       //The real map object
-      var map = new Map({
-       controls: [],
-        view: new View({
-          center:  fromLonLat((props.center.length==2 ? props.center : props.defaults.center) || geoLoc || []),
-          zoom: props.zoom || props.defaults.zoom ,
-          maxZoom: props.maxZoom || props.defaults.maxZoom || 1000, 
-          pitch : props.pitch || props.defaults.pitch,
-          rotation: props.rotation || props.defaults.rotation
-        }),
-        target: 'GEO_'+ geoId,
-        layers: [
-        //  baseLayer,
-          osm,
-         // geoJson,
-          vector,
-        ]
+      var olMap = new Map({
+        controls: [],
+         view: new View({
+           center:  fromLonLat((props.center.length==2 ? props.center : props.defaults.center) || geoLoc || []),
+           zoom: props.zoom || props.defaults.zoom ,
+           maxZoom: props.maxZoom || props.defaults.maxZoom || 100, 
+           rotation: props.rotation || props.defaults.rotation
+         }),
+         target: 'GEO_'+ geoId,
+         layers: []
+       });
+
+      // Click event listener for vector features and WMS GetFeatureInfo
+      olMap.on('singleclick', function (evt) {
+        let hasFeature = false;
+        fireEvent('click:single',evt)
+
+        olMap.forEachFeatureAtPixel(evt.pixel, function (feature, layer) {
+          // Vector feature click logic
+          hasFeature = true; // Indicate that a vector feature was clicked
+          fireEvent('click:feature',{properties: feature.getProperties(),layer})
+          return true; // Stop iterating through features
+        });
+
+        // WMS GetFeatureInfo logic
+        if (!hasFeature) { // Only proceed if no vector feature was clicked
+          olMap.getLayers().forEach(layer => {
+            if (layer instanceof TileLayer && layer.getSource() instanceof TileWMS) {
+              const view = olMap.getView();
+              const viewResolution = view.getResolution();
+              const url = layer.getSource().getFeatureInfoUrl(
+                evt.coordinate,
+                viewResolution,
+                'EPSG:3857',
+                { 'INFO_FORMAT': 'text/html' }, // or application/json ?
+              );
+              if (url) {
+                fetch(url)
+                  .then(response => response.text())
+                  .then(html => {
+                    // Display the HTML response in an element, or process JSON as needed
+                    fireEvent('click:reponse',{reponse: html})
+                    // Example: document.getElementById('info').innerHTML = html;
+                  });
+              }
+            }
+          });
+        }
       });
 
-      //Set the noMenu class based on flag
-      var menu = true;
-      //Set menuTitle and content
-      var menuTitle = "Menu"
-      var menuContent = `<p style="border-bottom:1px solid #999;">
-      <i>ol.control.Overlay</i> can be used to display a menu or information on the top of the map.
-    </p>`
-
-      // Menu Overlay
-      var menu = new Overlay ({ 
-        closeBox : true, 
-        className: "slide-left menu", 
-        content: `<div id="menuTitle"><h1>${menuTitle}</h1></div>
-        <div id="menuContent">${menuContent}</div>`
+      // Optional: pointer move logic for changing cursor over WMS layers
+      olMap.on('pointermove', function (evt) {
+        if (evt.dragging) return;
+        const pixel = olMap.getEventPixel(evt.originalEvent);
+        const hit = olMap.hasFeatureAtPixel(pixel);
+        olMap.getTargetElement().style.cursor = hit ? 'pointer' : '';
       });
-      if (!menu) menu.element.classList.add('nomenu')
-      map.addControl(menu);
-    
+
+      // Notification Control
+      olMap.addControl(notification);
+
+      //Handle the loaded event
+      olMap.on('loadend', function (event) {
+        fireEvent('loaded',event)
+      });
+
+      //Handle zoom event
+      olMap.getView().on('change:resolution', (event) => {
+        fireEvent('zoom',Object.assign({},event,{newValue: olMap.getView().getResolution()}))
+      });
+
+      //Add the buttons contols
       var zoom =  new Zoom({
         className:'ol-zoom',
         zoomInLabel: '+',
         zoomOutLabel: '-'
       })
-      if (!menu) zoom.element.classList.add('nomenu')
-      map.addControl(zoom)
-
-      // A toggle control to show/hide the menu
-      var toggle = new Toggle({
-        className: 'menu',
-        html: '<i class="fa fa-bars" ></i>',
-        title: "Menu",
-        onToggle: function() { menu.toggle(); }
-      });
-      if (menu) map.addControl(toggle);
-
+      if (!showButton('menu')) zoom.element.classList.add('nomenu')
+      if (showButton('zoom')) olMap.addControl(zoom)
 
       //Main menubar
       var mainbar = new Bar({className:"mainbar"});
       mainbar.setPosition("top-left")
-      if (!menu) mainbar.element.classList.add('nomenu')
-      map.addControl(mainbar);
+      if (!showButton('menu')) mainbar.element.classList.add('nomenu')
+      if (showButton('point')|| showButton('line') || showButton('polygon') ||
+      showButton('undo') || showButton('redo') || showButton('save'))  
+        olMap.addControl(mainbar);
 
       // Edit control bar 
       var editbar = new Bar({
@@ -205,17 +210,17 @@ function Geo(props) {
         title: 'Point',
         interaction: new Draw({
           type: 'Point',
-          source: vector.getSource()
+          source: drawVector.getSource()
         })
       });
-      editbar.addControl ( pedit );
+      if (showButton('point')) editbar.addControl ( pedit );
 
       var ledit = new Toggle({
         html: '<i class="fa fa-share-alt" ></i>',
         title: 'Line',
         interaction: new Draw({
           type: 'LineString',
-          source: vector.getSource(),
+          source: drawVector.getSource(),
           // Count inserted points
           geometryFunction: function(coordinates, geometry) {
             if (geometry) geometry.setCoordinates(coordinates);
@@ -225,7 +230,7 @@ function Geo(props) {
           }
         })
       });
-      editbar.addControl ( ledit );
+      if (showButton('line')) editbar.addControl ( ledit );
 
       var fedit = new Toggle({
         html: '<i class="fa fa-bookmark fa-rotate-270" ></i>',
@@ -233,7 +238,7 @@ function Geo(props) {
         interaction: new Draw({
           type: 'Polygon',
           style: [lightStroke, darkStroke],
-          source: vector.getSource(),
+          source: drawVector.getSource(),
           // Count inserted points
           geometryFunction: function(coordinates, geometry) {
             //this.nbpts = coordinates[0].length;
@@ -243,11 +248,19 @@ function Geo(props) {
           }
         })
       });
-      editbar.addControl ( fedit );
+      if (showButton('polygon')) editbar.addControl ( fedit );
 
-          // Undo redo interaction
-    var undoInteraction = new UndoRedo();
-    map.addInteraction(undoInteraction);
+      // Undo redo interaction
+      var undoInteraction = new UndoRedo();
+      undoInteraction.on('stack:add',function(e){
+        _setSkipRedraw(true)
+        fireEvent("draw:add", new GeoJSON().writeFeaturesObject(drawVector.getSource().getFeatures()))
+      })
+      undoInteraction.on('stack:remove',function(e){
+        _setSkipRedraw(true)
+        fireEvent("draw:remove", new GeoJSON().writeFeaturesObject(drawVector.getSource().getFeatures()))
+      })
+      olMap.addInteraction(undoInteraction);
 
       // Add a simple push button to undo features
       var undo = new Button({
@@ -257,98 +270,111 @@ function Geo(props) {
           undoInteraction.undo();
         }
       });
-      mainbar.addControl (undo );
-        // Add a simple push button to redo features
-        var redo = new Button({
+      if (showButton('undo')) mainbar.addControl (undo );
+      
+      // Add a simple push button to redo features
+      var redo = new Button({
         html: '<i class="fa fa-redo"></i>',
         title: "Redo",
         handleClick: function(e) {
           undoInteraction.redo();
         }
       });
-      mainbar.addControl (redo);
+      if (showButton('redo')) mainbar.addControl (redo);
 
       // Add a simple push button to save features
       var save = new Button({
         html: '<i class="fa fa-download"></i>',
         title: "Save",
         handleClick: function(e) {
-          var json= new GeoJSON().writeFeatures(vector.getSource().getFeatures());
-          console.log(json);
+          fireEvent("tracker:save",new GeoJSON().writeFeaturesObjects(trackerVector.getSource().getFeatures()))
         }
       });
-      mainbar.addControl (save );
+      if (showButton('save')) mainbar.addControl (save );
 
       //Fullscreen
       var fullscreen = new FullScreen()
-      map.addControl(fullscreen)
+      if (showButton('fullscreen')) olMap.addControl(fullscreen)
 
       var secondbar = new Bar();
       secondbar.setPosition("top-right")
-      map.addControl(secondbar);
+      if (showButton('layers')|| showButton('swipeHorizontal') ||showButton('swipeVertical') 
+        || showButton('timeline') ) olMap.addControl(secondbar);
 
       // Add a simple push button to save features
-      var layers = new Toggle({
+      var layersMenu = new Toggle({
         html: '<i class="fa fa-layer-group"></i>',
         title: "Layers",
-        handleClick: function(e) {
-          var json= new GeoJSON().writeFeatures(vector.getSource().getFeatures());
-          console.log(json);
+        onToggle: function(e) {
+          fireEvent("layers")
         }
       });
-      secondbar.addControl (layers );
+      if (showButton('layers')) secondbar.addControl (layersMenu );
 
-      /*
-      // Edit control bar 
-      var seditbar = new Bar({
+      // Swipe control bar 
+      var swipebar = new Bar({
         toggleOne: true,	// one control active at the same time
         group:false			// group controls together
       });
-      secondbar.addControl(seditbar);
-      */
+      if (showButton('swipeVertical') || showButton('swipeHorizontal'))
+          secondbar.addControl(swipebar);
 
-      var split = new Toggle({
-        html: '<i class="fa fa-table-columns"></i>', 
-        title: "Split",
-        handleClick: function(e) {
-          var json= new GeoJSON().writeFeatures(vector.getSource().getFeatures());
-          console.log(json);
+      var swipectrl = new Swipe({});
+      swipectrl.set('position',0,5)
+      //Todo Add the layers for the swipe control
+
+      var swipeHorz = new Toggle({
+        html: '<i class="fa fa-grip-lines-vertical"></i>', 
+        title: "Swipe Horizontal",
+        onToggle: function(event) {
+          if (event.active) {
+            swipectrl.set('orientation','horizontal')
+            olMap.addControl(swipectrl)
+          } else {
+            olMap.removeControl(swipectrl)
+          }
+          fireEvent("swipe:horizontal",event)
         }
       });
-      secondbar.addControl (split );
+      if (showButton('swipeHorizontal')) swipebar.addControl (swipeHorz );
 
-      var timeline = new Toggle({
-        html: '<i class="fa fa-clock"></i>', 
-        title: "Timeline",
-        onToggle: function(e) {
+      var swipeVert = new Toggle({
+        html: '<i class="fa fa-grip-lines-vertical fa-rotate-90"></i>', 
+        title: "Swipe Vertical",
+        onToggle: function(event) {
+          if (event.active) {
+            swipectrl.set('orientation','vertical')
+            olMap.addControl(swipectrl)
+          } else {
+            olMap.removeControl(swipectrl)
+          }
+          fireEvent("swipe:vertical",event)
         }
       });
-      secondbar.addControl (timeline );
+      if (showButton('swipeVertical')) swipebar.addControl (swipeVert );
 
 
-      //GeoLocation
-      var geoloc = new GeolocationButton({
-        title : "GeoLocation",
-        delay : 5000
+      // Menu Overlay
+      var menu = new Overlay ({ 
+        closeBox : true, 
+        className: "slide-left menu", 
+        content: `<div id="menuTitle" ><h1 id="menuTitle_`+geoId+`">${props.menuTitle.trim() || (props.defaults && props.defaults.menuTitle) || "&nbsp;"}</h1></div>
+        <div id="menuContent_`+geoId+`">${props.menuContent || (props.defaults && props.defaults.menuContent)}</div>`
       });
-      map.addControl(geoloc);
-      geoloc.on("change:active",(event)=>{
-        if (event.active) {
-          notification.show("Searching GPS",3000)
+      if (!showButton('menu')) menu.element.classList.add('nomenu')
+      olMap.addControl(menu);
+
+      // A toggle control to show/hide the menu
+      var toggle = new Toggle({
+        className: 'menu',
+        html: '<i class="fa fa-bars" ></i>',
+        title: "Menu",
+        onToggle: function(event) { 
+          menu.toggle(); 
+          fireEvent("menu:toggle",event)
         }
       });
-      //change:active
-
-    
-      //Add a GeoTracker
-      var geoTracker = new GeolocationBar({
-        source: vector.getSource(),
-        delay : 5000,
-        followTrack: 'auto',
-        minZoom: 16,
-        minAccuracy:10000
-      });
-      map.addControl(geoTracker)
+      if (showButton('menu')) olMap.addControl(toggle);
 
       var histo = [
         /* no more ?
@@ -360,7 +386,6 @@ function Geo(props) {
         }),
         */
       ]
-
       //Timeline
       var tline = new Timeline({
         className: 'ol-pointer ol-zoomhover ol-timeline',
@@ -390,58 +415,151 @@ function Geo(props) {
       tline.on('select', function(e) {
         tline.setDate(e.feature);
       });
-      
-      // CanvasScaleLine control
-      var scaleLineControl = new CanvasScaleLine();
-      map.addControl(scaleLineControl);
 
+      var timeline = new Toggle({
+        html: '<i class="fa fa-clock"></i>', 
+        title: "Timeline",
+        onToggle: function(e) {
+          fireEvent("timeline",e)
+        }
+      });
+      if (showButton('timeline')) secondbar.addControl (timeline );
       //Toggle the timeline classes
       timeline.on("change:active",(event)=>{
         if (event.active) {
           scaleLineControl.element.classList.add('timeline')
           geoTracker.element.classList.add('timeline')
           geoloc.element.classList.add('timeline')
-          map.addControl(tline);
+          olMap.addControl(tline);
+          fireEvent('timeline:active')
         } else {
           scaleLineControl.element.classList.remove('timeline')
           geoTracker.element.classList.remove('timeline')
           geoloc.element.classList.remove('timeline')
-          map.removeControl(tline);
+          olMap.removeControl(tline);
           //Work arround voor scaleLineControl not moving
-          map.removeControl(scaleLineControl);
-          map.addControl(scaleLineControl);
+          olMap.removeControl(scaleLineControl);
+          olMap.addControl(scaleLineControl);
+          fireEvent('timeline:inactive')
         }
       });
 
-      // Notification Control
-      var notification = new Notification({
+      //GeoLocation
+      var geoloc = new GeolocationButton({
+        title : "GeoLocation",
+        delay : 5000
       });
-      map.addControl(notification);
+      if (showButton('location')) olMap.addControl(geoloc);
+      geoloc.on("change:active",(event)=>{
+        if (event.active) {
+          notification.show("Searching GPS",3000)
+          fireEvent("geoloc:search")
+        }
+      });
+      //change:active
+    
+      //Add a GeoTracker
+      var geoTracker = new GeolocationBar({
+        source: trackerVector.getSource(),
+        delay : 5000,
+        followTrack: 'auto',
+        minZoom: 16,
+        minAccuracy:10000
+      });
+      if (showButton('tracker')) olMap.addControl(geoTracker)
+      var rotateNorth = new RotateNorthControl();
+      if (showButton('rotateNorth')) mainbar.addControl (rotateNorth );
 
-      //Add the supported events
-      map.on('loadend', function (event) {
-        props.onLoadEnd(event,notification)
-      });
-      map.on('click', function(event) {
-        props.onClick(event,notification)
-      });
-      map.getView().on('change:resolution', (event) => {
-        props.onZoom(event, map.getView().getResolution(),notification)
-      });
+      // CanvasScaleLine control
+      var scaleLineControl = new CanvasScaleLine();
+      if (showButton('scale')) olMap.addControl(scaleLineControl);
 
-
-      if (!props.showLogo) {
-        //geoCanvas.getElementsByClassName('l7-control-logo')[0].style="display:none" 
+      //On move
+      olMap.on('moveend', () => {
+        const extent = olMap.getView().calculateExtent(olMap.getSize()); // Get the current extent
+        const transformedExtent = transformExtent(extent, 'EPSG:3857', 'EPSG:4326'); // Transform the extent to WGS 84
+        fireEvent('bbox:change',transformedExtent); // Call the callback with the updated bbox
       }
-    }
-  }, [geoRef, props.center,props.zoom,props.maxZoom,props.rotation, props.pitch,props.geoJson,
-      props.showLogo,
-      props.onDataChange,props.onLoadEnd,props.onZoom,props.defaults,geoLoc
-  ]);
+     );
 
+      setMap(olMap)
+      }
+    }, [geoRef,props.defaults, props.buttons]);  
+
+    //Zoom handling
+    useEffect(() => {
+      if (map) map.getView().setZoom(props.zoom)
+    },[props.zoom]);
+    //Max zoom handling
+    useEffect(() => {
+      if (map) {
+        map.getView().setMaxZoom(props.maxZoom)
+        map.getView().setZoom(Math.min(props.zoom,props.maxZoom))
+      }
+    },[props.maxZoom]);
+    //rotation handling
+    useEffect(() => {
+      if (map) {
+        map.getView().setRotation(props.rotation)
+      }
+    },[props.rotation]);
+    //Center handling
+    useEffect(() => {
+      if (map && props.center && props.center.length==2) map.getView().setCenter(props.center)
+    },[props.center]);
+    //Menu title
+    useEffect(() => {
+      if (map) {
+        var el=document.getElementById('menuTitle_'+geoId)
+        if (el) el.innerHTML=(props.menuTitle.trim() || (props.defaults && props.defaults.menuTitle) || "&nbsp;")
+      }
+    },[props.menuTitle]);
+    //Menu content
+    useEffect(() => {
+      if (map) {
+        var el=document.getElementById('menuContent_'+geoId)
+        if (el) el.innerHTML=props.menuContent || (props.defaults && props.defaults.menuContent)
+      }
+    },[props.menuContent]);
+
+     // Dynamic layer updating
+    useEffect(() => {
+      if (map && allowUpdate('drawLayer')) {
+        // Validate and create new layers
+        const layers = Array.isArray(props.layers) ? props.layers : 
+            props.defaults && Array.isArray(props.defaults.layers) ? props.defaults.layers : [];
+        const validatedLayers = layers.filter(layer => layer !== null && layer !== undefined);
+        const sortedLayers = validatedLayers
+          .sort((a, b) => (a.order || 0) - (b.order || 0))
+          .map(createLayer)
+          .filter(layer => layer !== null && layer !== undefined);
+        map.getLayers().clear();
+        sortedLayers.forEach(layer => map.addLayer(layer));
+        //ToDo connect to the switch controler, add draw and tracker layer
+
+        //Add drawLayer and values if set
+        var drawSource = new VectorSource()
+        drawVector.setSource(drawSource)
+        map.addLayer(drawVector)
+        if (props.drawLayer) {
+          try {
+            var geojsonFormat = new GeoJSON();
+            // reads and converts GeoJSon to Feature Object
+            var features = geojsonFormat.readFeatures(props.drawLayer);
+            drawSource.addFeatures(features)
+          } catch(e){
+            if (props.debug) console.log("drawLayer invalid json")
+          }
+        }
+
+        //Trackerlayer
+        map.addLayer(trackerVector)
+        fireEvent("layers:update",layers)
+      }
+    }, [map, props.layers, props.drawLayer]); // Re-evaluate when layers change
   return (
     <div
-      ref={useGeoRef}
+      ref={elementRef}
       style={{ height: '100%', width: '100%' }}
     >
     <RingLoader color="#36d7b7" 
@@ -457,14 +575,14 @@ Geo.propTypes = {
   zoom: PropTypes.number,
   maxZoom: PropTypes.number,
   rotation: PropTypes.number,
-  pitch: PropTypes.number,
-  geoJson: PropTypes.object,
-  showLogo : PropTypes.bool,
-  onDataChange: PropTypes.func,
-  onLoadEnd: PropTypes.func,
-  onClick: PropTypes.func,
-  onZoom: PropTypes.func,
+  drawLayer: PropTypes.object,
+  onEvent : PropTypes.func,
   skipRedraw: PropTypes.func,
+  buttons: PropTypes.object,
+  menuTitle: PropTypes.string,
+  menuContent: PropTypes.string,
+  ignoreUpdate: PropTypes.func,
+  layers: PropTypes.array,
   defaults: PropTypes.object,
 }